requirements.txt
requests
pyyaml
logging